Changed intval to floatval
git-svn-id: https://192.168.1.50:9880/easyshop_svn/trunk@3055 2929c1f8-612a-3548-b275-2a983e7c7391

+++ application/models/product_model.php
@@ -2035,7 +2035,8 @@ class product_model extends CI_Model
             $result['is_soldout'] = $this->is_sold_out($product_id);
             $result['sold_price'] = $this->get_sold_price($product_id, date('Y-m-d',$startdate), date('Y-m-d',$enddate));
         }
-        $result['price'] = (intval($result['price'])>0)?$result['price']:0.01;
+        
+        $result['price'] = (floatval($result['price'])>0)?$result['price']:0.01;
         return $result;
    }
